isPrecompiledContract() moved into ContractInfo
- This method doesn't belong in a factory class, just separating it out.
