+++ modPrecompiled/src/org/aion/precompiled/ContractInfo.java
@@ -26,4 +26,20 @@ public enum ContractInfo {
         this.ownerAddress =
                 (owner == null) ? null : new AionAddress(ByteUtil.hexStringToBytes(owner));
     }
+
+    /**
+     * Returns true if address is the address of a pre-compiled contract and false otherwise.
+     *
+     * @param address The address to check.
+     * @return true iff address is address of a pre-compiled contract.
+     */
+    public static boolean isPrecompiledContract(AionAddress address) {
+        for (ContractInfo contractInfo : ContractInfo.values()) {
+            if ((contractInfo != ContractInfo.TOTAL_CURRENCY)
+                    && (address.equals(contractInfo.contractAddress))) {
+                return true;
+            }
+        }
+        return false;
+    }
 }
